requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Pydantic-Spec==0.11.1

Gunicorn==21.2.0

requests==2.31.0

pydantic-settings==2.6.0
python-dotenv==1.0.0

logging==0.5.1.2

requests==2.31.0